backend/requirements.txt
# Django and REST Framework
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.0
PyJWT==2.8.0

# Google APIs
google-api-python-client==2.114.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Server
gunicorn==21.2.0
whitenoise==6.6.0

# Caching
redis==5.0.1
django-redis==5.4.0

# Utilities
python-dotenv==1.0.0
Pillow==10.2.0
celery==5.3.4

# Development and Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.4.0

# API Documentation
drf-spectacular==0.27.0
